split functions and accidentally made submission
didn't mean to make the submission within this branch, but I completed the functionality I wanted and appropriately split apart the functionalities into their own middleware like this should have been, instead of that giant "store" file.
