+++ src/middleware/index.ts
@@ -1,3 +1,6 @@
 import * as store from './store';
 import * as auth from './authenticate'
-export {store, auth};
+import * as user from './user'
+import * as host from './host'
+import * as prop from './propagandist'
+export {store, auth, user, host, prop};
